requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
line-bot-sdk==2.4.2
azure-keyvault-secrets
azure-identity
bardapi~=0.1.24
requests~=2.28.2
beautifulsoup4~=4.12.2